requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
python-multipart>=0.0.9
pyyaml>=6.0.1
jinja2>=3.1.3
markdown>=3.6
